Change sample tests.
